Add authentication to controllers.
Fixed adding created_at and updated_at in reservations_seats table.

Add getReservationSeatsWithMovie function in ReservationController and route in api.php.

+++ app/Http/Controllers/HomePageSlideController.php
@@ -31,6 +31,11 @@ class HomePageSlideController extends Controller
     }
     public function delete($id)
     {
+        $user = Auth::user();
+
+        if (!$user || $user->role !== 'admin') {
+            return response()->json(['error' => 'Forbidden'], 403);
+        }
         $slide = HomePageSlide::findOrFail($id);
         $slide->delete();
         return response()->json(['message' => 'Slide deleted successfully']);
